requirements.txt
requests
moviepy
Pillow
numpy
google-generativeai
